Articles grouping.

+++ Paperboy/Views/FeedItemListRow.swift
@@ -14,7 +14,7 @@ struct FeedItemListRow: View {
     
     var dateFormatter: DateFormatter {
         let formatter = DateFormatter()
-        formatter.dateStyle = .long
+        formatter.dateStyle = .short
         formatter.timeStyle = .short
         return formatter
     }
